+++ .github/changeLog.md
@@ -1,10 +1,17 @@
-# Chnagelog 
+# Changelog 
 
-The verison 1.0 will be the actual usuable release.
+The version 1.0 will be the actual usable release.
 
 ---
+# Version 0.0.2
 
-# Version 0.1 
+Added:
+- Connected to a Database
+- Events handler
+- Command handler
+- Ping Command
+
+# Version 0.0.1 
 
 Added:
 - Index.js